Deployment/requirements.txt
fastapi==0.89.1
nltk==3.7
numpy==1.23.5
pydantic==1.10.2
python-dotenv==0.21.1
tensorflow==2.10.0
uvicorn==0.20.0